requirements.txt
alembic==1.15.2
altair==5.5.0
asttokens==2.4.1
attrs==25.1.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.2
catboost==1.2.7
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.0
coverage==7.6.7
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
executing==2.1.0
fonttools==4.54.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
greenlet==3.1.1
hyperopt==0.2.7
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-black==0.4.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
narwhals==1.29.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
optuna==4.2.1
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
pprintpp==0.4.0
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.0.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==19.0.1
pycountry==24.6.1
pycountry-convert==0.7.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
repoze.lru==0.7
requests==2.32.3
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.8.0
shap==0.47.1
six==1.16.0
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
SQLAlchemy==2.0.40
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.43.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tokenize_rt==6.1.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
wheel==0.45.1
xgboost==2.1.4
